// src/app/shared/models/finance-Pro/daily-collection-report.ts
export class DailyCollectionReport {
    academicSessionId?: number;
    academicSessionName?: string;
    programId?: number;
    programName?: string;
    operationalVerticalId?: number;
    operationalVerticalName?: string;
    feeComponentId?:number;
    feeComponentName?:string;
    studentId?: string;
    registrationNumber?: string;
    studentName?: string;
    receiptDate?: string;
    paymentDate?: string;
    receiptNumber?: string;
    paidAmount?: number;
    referenceNumber?: string;
    paymentMode?: number;
    paymentModeType?: number;
    remarks?: string;
    status?: string;
}